Added the BSD licensed xmlformat utility for keeping the xml pretty and consistant aswell as a configuration file. When I converted the manual to docbook I was using an old version that I'd modified to insert tabs rather than spaces so using it on the existing documentation will appear to reformat the whole file. I don't think that matters at this stage and I'm tempted to reformat everything now while it isn't an issue.
Rewrote the Makefile that builds the docs so that:
- building the html only depends on having xsltproc installed.
- the xsl and xml files aren't copied to the build directory.
- make test does not remove an existing build of the docs.

Changed the glossary so that the acronym/abbreviations are the 
glossary terms, which I've discovered is how it is generally done.
It makes total sense now that I actually think about it because
you need to find the terms based on their use in the manual.

Added a bit of xsl to add the status attribute as a class attribute 
in the html so that the css can use a draft watermark(or any other 
watermark). DocBook supports the draft status without the xsl 
but it has problems, which I've described in the xsl file.

Minor tweak to the css to make the glossary terms and any other 
definition lists bold(that includes the index).


git-svn-id: http://subversion.ardour.org/svn/ardour2/trunk@1457 d708f5d6-7413-0410-9779-e7cbd77b26cf