The Big Change: Store time in MidiModel as tempo time, not frame time. The time stamp of an event is now always tempo, from file to model and back again.  Frame time is only relevant at playback or recording time, in the audio thread (MidiModel and MidiBuffer).
I think perhaps we don't need to change the actual time from double (which is
convenient for math), it is the time base conversion that caused problems.
Using a correct equality comparison (i.e.  not == which is not correct for
floating point) should probably make the undo issues go away, in 99.99% of
cases anyway.

There's almost certainly some regressions in here somewhere, but they do not
seem to be time related.  The bugs I'm hitting in testing are old ones that
seem unrelated now, so it's checkpoint time.

This sets us up for fancy things like tempo map import and tempo/meter changes
halfway through MIDI regions, but for now it's still assumed that the tempo
at the start of the region is valid for the duration of the entire region.


git-svn-id: http://subversion.ardour.org/svn/ardour2/branches/3.0@4582 d708f5d6-7413-0410-9779-e7cbd77b26cf