Introduce new time for session-relative frame time, and make source interface capable of handling 64-bit long sessions. sframes_t is "session frames".  The rules for time stamps are:  - Anything relative to transport time, session position, etc, should be sframes_t  - Anything relative to jack cycles, including the length thereof, should be nframes_t
To support sessions which exceed UINT32_MAX frames, we need to replace all the uses of
nframes_t for session time with sframes_t, and make sure the conversions are sound.
This does not depend on jack's nframes_t; that we are using the same type at all right now was an oops.

This is also be kinda nice for readability since the two different time bases have different types...


git-svn-id: http://subversion.ardour.org/svn/ardour2/branches/3.0@4636 d708f5d6-7413-0410-9779-e7cbd77b26cf